backend/requirements.txt
Django==5.2.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.3.1
psycopg2==2.9.10
python-dotenv==1.0.1
pillow==11.1.0
PyJWT==2.9.0
sqlparse==0.5.2
asgiref==3.8.1
